LayoutTests/http/tests/navigation/https-no-store-subframe-in-page-cache.html

<script>

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
    testRunner.overridePreference("WebKitUsesPageCachePreferenceKey", 1);
}

if (location.protocol != "https:")
    location = "https://127.0.0.1:8443/navigation/resources/https-no-store-subframe-in-page-cache.html";

</script>
