LayoutTests/http/tests/navigation/https-no-store-subframe-in-page-cache-expected.txt

ALERT: This page is https and has has a subframe with the no-store cache-control directive. It should go in to the page cache.
ALERT: PASS: The page was restored from the page cache. Good job!
